Refactor ambassador validation to conform to DID lib Validator interface
